refactor: rename module name
